Add some test coverage of setter methods with validation errors.
Existing specs against setter methods are changed so that they're
working with a persisted record. Because the record is now persisted, we
can correctly `reload` the record to verify that, e.g., `empty!` really
does persist the flexible enum value, and its related timestamps.
Without this assertion, it's possible that `update_attributes()` is
actually failing to persist the new values and is only updating the
value in memory.

Note that the previous wording of "validation-free" was mistaken:
`update_attributes()` (plural) does not skip validations (as it's an
alias for `update()`. (`update_attribute()`, however, does skip
validations.)