Have setter methods raise exceptions.
Consider the following flexible enum configuration.

  class User < ActiveRecord::Base
    flexible_enum :status do
      active 1
      inactive 2, setter: :deactivate!
    end
  end

When a `User` is deactivated, it's likely that you'd see something like:

  if deactivation_requested
    current_user.deactivate!
  end

Because of the naming of the method (with a bang) and existing Rails
conventions, it's unlikely that you'd check the result of
`current_user.deactivate!` to see if it was successful.

Moreover, there are situations where you wish to clear out related
records for all deactivations. You can currently do that by wrapping the
setter:

  def deactivate!
    transaction do
      self.foos.each(&:remove!)
      self.bars.each(&:remove!)

      super
    end
  end

An expectation with this code is that `super` will `raise` if there are
any problems, rolling back the transaction. Except what actually would
happen currently is that `super` would return `false` (through
`update_attributes()`, no `!`), the transaction would commit, and now
the user is still `ACTIVE` but the `foos` and `bars` were removed.

This changes all setter methods to instead `raise` when there are
problems with the update.